test(appsec): move telemetry tests to product suites
